+++ src/components/layout/header/HeaderStyles.ts
@@ -17,6 +17,33 @@ export const Logo = styled.img`
   height: 80px;
 `;
 
+export const SearchBox = styled.div`
+  display: flex;
+  align-items: center;
+  position: absolute;
+  top: 50%;
+  right: 10%;
+  transform: translate(0, -50%);
+  border: 1px solid black;
+  border-radius: 20px;
+  overflow: hidden;
+`
+
+export const SearchImage = styled.img`
+  width: 30px;
+  padding-left: 5px;
+
+`
+
+export const Search = styled.input`
+  width: 150px;
+  display: block;
+  padding: 5px;
+  border: none;
+  outline: none;
+  font-size: 1.2rem;
+`
+
 export const IconBox = styled.div`
   position: absolute;
   top: 50%;
@@ -43,6 +70,7 @@ export const Menu = styled.ul`
   padding: 0;
   margin: 0;
   font-size: 1.2rem;
+  z-index: 999;
 `;
 
 export const MenuItem = styled.li`
